Extract algorithm tests from ring_algs.rs
