+++ Cargo.toml
@@ -45,6 +45,7 @@ include = [
     "src/name/name.rs",
     "src/signed_data.rs",
     "src/ring_algs.rs",
+    "src/alg_tests.rs",
     "src/time.rs",
     "src/trust_anchor.rs",
     "src/x509.rs",
